Vertical Slice

+++ Assets/HealthBar.cs
@@ -7,12 +7,12 @@ public class HealthBar : MonoBehaviour
 {
     public Slider healthSlider;
 
-    public void SetSlider(float amount)
+    public void SetSlider(int amount)
     {
         healthSlider.value = amount;
     }
 
-    public void SetSliderMax(float amount)
+    public void SetSliderMax(int amount)
     {
         healthSlider.maxValue = amount;
         SetSlider(amount);
